refeat: split JSX into separated components
