+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { ErrorBoundary } from './components/error/ErrorBoundary';
+import { ErrorBoundary } from './pages/ErrorBoundary';
 import catImage from './assets/cat.webp';
 import { Provider } from 'react-redux';
 import { store } from './store';
