+++ src/pages/about/components/BackButton.tsx
@@ -1,0 +1,9 @@
+import { Link } from 'react-router-dom';
+
+export function BackButton() {
+  return (
+    <Link to="/" className="back-button">
+      ← Back to Pokémon
+    </Link>
+  );
+}
