+++ src/pages/pokemon/components/detail/PokemonDetail.css
