+++ src/pages/pokemon/components/detail/components/PokemonAbilities.tsx
@@ -1,0 +1,16 @@
+import { PokemonDetails } from '@/store/types/pokemon';
+interface PokemonAbilitiesProps {
+  pokemon: Pick<PokemonDetails, 'abilities'>;
+}
+export function PokemonAbilities({ pokemon }: PokemonAbilitiesProps) {
+  return (
+    <div className="pokemon-abilities">
+      <h4>Abilities</h4>
+      <ul>
+        {pokemon.abilities?.map((ability) => (
+          <li key={ability}>{ability.replace('-', ' ')}</li>
+        ))}
+      </ul>
+    </div>
+  );
+}
