+++ src/pages/pokemon/components/list/PokemonList.css
