+++ src/pages/pokemon/components/panel/SelectedPokemonsPanel.css
